requirements.txt
aiohttp==3.6.3
arduino-sketch==0.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
discord.py==1.5.1
dnspython==2.0.0
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.2
opencv-contrib-python==4.4.0.44
openpyxl==3.0.5
pandas==1.1.3
Pillow==8.0.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
six==1.15.0
tabulate==0.8.7
tk==0.1.0
toml==0.10.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.3.7
yarl==1.5.1
